[WIP] Start moving notification builder to new module
