+++ components/bsb/timer/notification/android/build.gradle
@@ -1,0 +1,15 @@
+plugins {
+    id("flipper.multiplatform")
+    id("flipper.multiplatform-dependencies")
+    id("flipper.anvil-multiplatform")
+}
+
+androidDependencies {
+    implementation(projects.components.bsb.timer.notification.api)
+
+
+    implementation(projects.components.bsb.timer.background.api)
+    implementation(projects.components.bsb.timer.notification.common)
+
+    implementation(libs.androidx.annotation)
+}
