+++ components/bsb/timer/notification/android/src/androidMain/kotlin/com/flipperdevices/bsb/timer/notification/NotificationTimerBuilderImpl.kt
@@ -1,0 +1,16 @@
+package com.flipperdevices.bsb.timer.notification
+
+import android.app.Notification
+import android.content.Context
+import com.flipperdevices.bsb.timer.background.model.ControlledTimerState
+
+class NotificationTimerBuilderImpl {
+    fun buildStartUpNotification(
+        context: Context
+    ): Notification
+
+    fun buildNotification(
+        context: Context,
+        timer: ControlledTimerState
+    ): Notification?
+}
