+++ components/bsb/timer/notification/android/src/androidMain/res/drawable/pic_status_rest_paused.xml
