+++ components/bsb/timer/notification/android/src/androidMain/res/values/styles.xml
