+++ components/bsb/timer/notification/api/src/androidMain/kotlin/com/flipperdevices/bsb/timer/notification/NotificationTimerBuilder.kt
@@ -1,0 +1,16 @@
+package com.flipperdevices.bsb.timer.notification
+
+import android.app.Notification
+import android.content.Context
+import com.flipperdevices.bsb.timer.background.model.ControlledTimerState
+
+interface NotificationTimerBuilder {
+    fun buildStartUpNotification(
+        context: Context
+    ): Notification
+
+    fun buildNotification(
+        context: Context,
+        timer: ControlledTimerState
+    ): Notification?
+}
