+++ components/bsb/timer/notification/common/build.gradle
@@ -1,0 +1,7 @@
+plugins {
+    id("flipper.multiplatform")
+    id("flipper.multiplatform-dependencies")
+}
+
+androidDependencies {
+}
