+++ components/bsb/timer/notification/wearos/build.gradle
@@ -1,0 +1,10 @@
+plugins {
+    id("flipper.multiplatform")
+    id("flipper.multiplatform-dependencies")
+}
+
+androidDependencies {
+    implementation(projects.components.bsb.timer.notification.api)
+
+    implementation(projects.components.bsb.timer.notification.common)
+}
